+++ apps/client/lib/l10n/app_localizations_en.dart
@@ -2095,6 +2095,15 @@ class AppLocalizationsEn extends AppLocalizations {
   String get resumeDownload => 'Resume';
 
   @override
+  String get updateDownloadingTitle => 'Downloading update';
+
+  @override
+  String get updateReadyTitle => 'Update ready';
+
+  @override
+  String get updateReadyBody => 'Tap to install';
+
+  @override
   String get addLink => 'Add link';
 
   @override